Fix parsing of values
The Bondora reports are dependent on the set locale. In some cases the
decimal separator is a `,` in other cases a `.`. The thousand separator
is reversed. Unfortunately Lua doesn't have any nice to convert a string
to a number while automatically take the locale into consideration.

The solution always requests the English representation and removes any
thousand separator while keeping the decimal separator.